% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.rqa_result}
\alias{autoplot.rqa_result}
\title{Recurrence plot for an \code{rqa_result}}
\usage{
\method{autoplot}{rqa_result}(object, ...)
}
\arguments{
\item{object}{An \code{rqa_result} from \code{\link[=rqa]{rqa()}}.}

\item{...}{Forwarded to \code{\link[=plot_recurrence]{plot_recurrence()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
S3 method for the \code{ggplot2::autoplot()} generic. Equivalent to
calling \code{\link[=plot_recurrence]{plot_recurrence()}} on the same object.
}
