% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqaem-package.R
\docType{package}
\name{rqaem-package}
\alias{rqaem}
\alias{rqaem-package}
\title{rqaem: Recurrence Quantification Analysis of Eye Movements}
\description{
Tools for recurrence quantification analysis (RQA) of fixation sequences as described in Anderson, Bischof, Laidlaw, Risko and Kingstone (2013) \doi{10.3758/s13428-012-0299-5}. Provides a tidy data-frame workflow with sparse recurrence matrices, bootstrap significance testing, and radius selection. Ports the reference Python and MATLAB implementations at bit-for-bit parity on the bundled test fixtures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nccanderson/rqaem}
  \item Report bugs at \url{https://github.com/nccanderson/rqaem/issues}
}

}
\author{
\strong{Maintainer}: Nicki Anderson \email{nicki@cognitive3d.com}

Other contributors:
\itemize{
  \item Walter F. Bischof (Author of the reference Python and MATLAB implementations) [contributor]
}

}
\keyword{internal}
